[SW-XXX] Preparation for scoring package
Extract methods which can be shared between core and scoring package to
separate utils submodule
