+++ build.gradle
@@ -42,7 +42,9 @@ ext {
             project(':sparkling-water-examples'),
             project(':sparkling-water-ml'),
             project(':sparkling-water-package'),
-            project(":sparkling-water-doc")
+            project(":sparkling-water-doc"),
+            project(':sparkling-water-scoring'),
+            project(':sparkling-water-utils')
     ]
     // Projects with integration tests
     integTestProjects = [
@@ -65,7 +67,9 @@ ext {
             project(':sparkling-water-app-streaming'),
             project(':sparkling-water-package'),
             project(':sparkling-water-benchmarks'),
-            project(':sparkling-water-macros')
+            project(':sparkling-water-macros'),
+            project(':sparkling-water-scoring'),
+            project(':sparkling-water-utils')
     ]
     javaProjects = [
             project(':sparkling-water-extension-stack-trace')
