+++ settings.gradle
@@ -18,6 +18,8 @@ include 'benchmarks'
 include 'dist'
 include 'macros'
 include 'jenkins'
+include 'scoring'
+include 'utils'
 
 // Prefix all projects with sparkling-water name
 rootProject.children.each { project ->
