+++ utils/build.gradle
@@ -1,0 +1,12 @@
+description = "Utilities shared accross Sparkling Water sub-modules. These utilities does not depend on H2O."
+
+apply from: "$rootDir/gradle/sparkTest.gradle"
+apply from: "$rootDir/gradle/utils.gradle"
+
+dependencies {
+    compileOnly "org.apache.spark:spark-core_${scalaBaseVersion}:${sparkVersion}"
+    compileOnly "org.apache.spark:spark-mllib_${scalaBaseVersion}:${sparkVersion}"
+}
+
+
+defineStandardPublication().call()
